requirements.txt
numpy==1.20.3
pandas==1.3.4
Pillow==9.1.0
streamlit==1.8.1
tensorflow==2.8.0
